+++ index.js
@@ -8,7 +8,7 @@ module.exports = {
   get expectEvent () { return require('./src/expectEvent'); },
   get makeInterfaceId () { return require('./src/makeInterfaceId'); },
   get send () { return require('./src/send'); },
-  get shouldFail () { return require('./src/shouldFail'); },
+  get expectRevert () { return require('./src/expectRevert'); },
   get singletons () { return require('./src/singletons'); },
   get time () { return require('./src/time'); },
 };